[Docker] Implement feedback round 3 for version badge
