+++ services/docker/docker-helpers.js
@@ -1,23 +1,9 @@
 'use strict'
 
 const dockerBlue = '066da5' // see https://github.com/badges/shields/pull/1690
-const Joi = require('@hapi/joi')
-const { latest } = require('../version')
 
-const queryParamSchema = Joi.object({
-  sort: Joi.string()
-    .valid('date', 'semver')
-    .default('date'),
-}).required()
-
-function buildDockerUrl(badgeName, includeTagRoute, includeQueryParam) {
-  if (includeTagRoute && includeQueryParam) {
-    return {
-      base: `docker/${badgeName}`,
-      pattern: ':user/:repo/:tag*',
-      queryParamSchema,
-    }
-  } else if (includeTagRoute && !includeQueryParam) {
+function buildDockerUrl(badgeName, includeTagRoute) {
+  if (includeTagRoute) {
     return {
       base: `docker/${badgeName}`,
       pattern: ':user/:repo/:tag*',
@@ -54,7 +40,13 @@ function getDigestSemVerMatches({ data, digest }) {
   const matches = data
     .filter(d => d.images.some(i => i.digest === digest))
     .map(d => d.name)
-  return latest(matches)
+  let version = matches[0]
+  matches.forEach(name => {
+    const dots = (name.match(/\./g) || []).length
+    const olddots = (version.match(/\./g) || []).length
+    version = dots >= olddots && name !== 'latest' ? name : version
+  })
+  return version
 }
 
 module.exports = {
